Game Session initial values are now loaded from an xml file so Designers can tune them without the need of a coder.

+++ Assets/02_Scripts/game/serialization/GameSessionData.cs
@@ -8,7 +8,7 @@ public class GameSessionData : GameData
 	public int growthRate;
 	public int capacity;
 	public int money;
-	public float publicOpinion;
+	public int publicOpinion;
 	public int[] patients;
 	public int[] advisorIds;
 }
